This should fix server issues...
The problem was that vapor always sees the local hostname as baseURL,
which messes with things if you put it behind a proxy. This has been
fixed, if the proxy provides a X-base-URL header to the vapor app. This
header must contain the baseURL to use for the app
(https://clocker.goranche.net)
This can be achieved on nginx with the following directive:
  proxy_set_header X-base-URL $scheme://$host;