Changed the instantiation code for the Droplet
The previous code used a deprecated constructor method. A memory DB
provider needed to be added, but this now causes a log entry about a
deprecated vapor API thingy... will deal with that when the time comes.
If the memory DB provider gets updated this log entry should disappear,
though... hopefully 😅